requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
python-multipart==0.0.9
python-dotenv==1.0.1

# Database and Security
sqlalchemy==2.0.28
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Storage and Cryptography
cryptography==42.0.5
minio==7.2.2

# Utilities
loguru==0.7.2
